refactor(directory): organizes components by page
