+++ app/components/HomePage/CompanyValues.tsx
@@ -15,8 +15,8 @@ import Image from "next/image";
 import { FaRegClock, FaPeopleGroup, FaRegStar } from "react-icons/fa6";
 
 // LOCAL IMPORTS
-import gearBackground from "../../public/images/machine-gear-background.png";
-import ContentWrapper from "./ContentWrapper";
+import gearBackground from "../../../public/images/machine-gear-background.png";
+import ContentWrapper from "../ContentWrapper";
 
 // VALUES COMPONENT DEFINITION
 export default function CompanyValues(props: PageBlocksValues) {
