+++ app/components/HomePage/Location.tsx
@@ -1,6 +1,6 @@
 import { Box, Text } from "@chakra-ui/react";
 import { PageBlocksLocation } from "@/tina/__generated__/types";
-import ContentWrapper from "./ContentWrapper";
+import ContentWrapper from "../ContentWrapper";
 
 export default function Location(props: PageBlocksLocation) {
   return (
