+++ app/components/HomePage/MachineCarousel.tsx
@@ -3,8 +3,8 @@ import React, { useState } from "react";
 import { ChevronLeftIcon, ChevronRightIcon } from "@chakra-ui/icons";
 import Slider from "react-slick";
 
-import { MachineCardsBlock } from "./Machines";
-import Card from "./Card/Card";
+import { MachineCardsBlock } from "../MachinesPage/Machines";
+import Card from "../Card/Card";
 
 type MouseEvent = React.MouseEventHandler<SVGElement>;
